requirements.txt
openai
flask
flask-login
flask-sqlalchemy
pandas
gunicorn